Benchmark:
old:
```
$ SEED=52d036450e91515f807900d2f2f73fc4194f3f0d602283f30509e41b4b5c472e node benchmark/benchmark.js
Seed: 52d036450e91515f807900d2f2f73fc4194f3f0d602283f30509e41b4b5c472e
--------------------------------------------------
encode x 51,640 ops/sec ±1.31% (9 runs sampled)
decode x 76,285 ops/sec ±0.45% (9 runs sampled)
==================================================
```

new:
```
$ SEED=52d036450e91515f807900d2f2f73fc4194f3f0d602283f30509e41b4b5c472e node benchmark/benchmark.js
Seed: 52d036450e91515f807900d2f2f73fc4194f3f0d602283f30509e41b4b5c472e
--------------------------------------------------
encode x 65,623 ops/sec ±1.06% (8 runs sampled)
decode x 287,033 ops/sec ±0.31% (9 runs sampled)
==================================================
```